authn-service/requirements.txt
Flask==1.1.2
Flask-Cors==3.0.8
PyJWT==1.7.1
requests==2.24.0